refactor categories route & controllers

+++ src/controllers/categoriesController.ts
@@ -11,45 +11,47 @@ export const createCategory = async (req: Request, res: Response) => {
   const { name } = req.body;
 
   if (!name) {
-    return res.status(400).json({ message: "Name is required" });
+    return res.status(400).json({ message: "Name is required!" });
   }
 
   const categoryExists = await categoryRepository.findOne({
     where: { name: ILike(name) },
   });
 
-  if (categoryExists) {
+  if (categoryExists)
     return res.status(400).json({ message: "Category already exists" });
-  } else {
-    const newCategory = new Category();
-    newCategory.name = name;
 
-    const newSubCategory = new SubCategory();
-    newSubCategory.name = name;
-    newSubCategory.category = newCategory;
+  const newCategory = new Category();
+  newCategory.name = name;
 
-    try {
-      await categoryRepository.save(newCategory);
-      await subCategoryRepository.save(newSubCategory);
+  //Think about the need of creating a general subcategory
+  const newSubCategory = new SubCategory();
+  newSubCategory.name = name;
+  newSubCategory.category = newCategory;
 
-      return res.status(201).json({
-        category: newCategory,
-        message: "Category created with success",
-      });
-    } catch (error) {
-      return res.status(500).json({ message: "Failed to create category" });
-    }
+  try {
+    await categoryRepository.save(newCategory);
+    await subCategoryRepository.save(newSubCategory);
+
+    return res.status(201).json({
+      category: newCategory,
+      message: "Category created with success!",
+    });
+  } catch (error) {
+    return res.status(500).json({ message: "Failed to create category" });
   }
 };
 
 export const getCategories = async (req: Request, res: Response) => {
   try {
-    const categories = await categoryRepository.find();
+    const categories = await categoryRepository.find({
+      relations: ["subcategories"],
+    });
 
     if (categories.length !== 0) {
       return res.status(200).send(categories);
     } else {
-      return res.status(400).send("No categories created");
+      return res.status(400).send("No categories created!");
     }
   } catch (error) {
     return res.status(500).json({ message: "Failed to get categories" });
@@ -58,32 +60,29 @@ export const getCategories = async (req: Request, res: Response) => {
 
 export const deleteCategory = async (req: Request, res: Response) => {
   const { id } = req.params;
+  const parsedId = Number(id);
 
   try {
     //Find the category to be deleted
-    const categoryToDelete = await categoryRepository.findOneByOrFail({
-      id: Number(id),
+    const categoryToDelete = await categoryRepository.findOne({
+      where: { id: parsedId },
     });
 
-    //Find subcategories associated to the category
-    const relatedSubcategories = await subCategoryRepository.find({
-      where: { category: categoryToDelete },
-    });
-
-    //Delete all subcategories
-    await subCategoryRepository.remove(relatedSubcategories);
+    if (!categoryToDelete) {
+      return res
+        .status(404)
+        .json({ message: `Category with id ${parsedId} not found!` });
+    }
 
     //Delete the category
     await categoryRepository.remove(categoryToDelete);
 
-    const { name } = categoryToDelete;
-    
-    return res
-      .status(200)
-      .json({ message: `Category ${name} successfully deleted!` });
+    return res.status(200).json({
+      message: `Category ${categoryToDelete.name} successfully deleted!`,
+    });
   } catch (error) {
     return res
       .status(404)
-      .json({ message: `Category with id ${id} not found` });
+      .json({ message: `Something went wrong while deleting the category!` });
   }
 };
